shared/idmap: Introduce set_sort
